+++ AutomatonTest.cpp
@@ -4,12 +4,12 @@
 typedef Automaton<bool>::Run   Run;
 typedef Automaton<bool>::Input Input;
 
-void test(Automaton<bool> &a, Input input[], int input_size)
+void test(Automaton<bool> &a, Input input[], size_t input_size)
 {
     bool accepted = a.Accept(input, input_size);
 
     printf("%s: ", (accepted ? "Accepted" : "Rejected"));
-    for (int i = 0; i < input_size; i++)
+    for (size_t i = 0; i < input_size; i++)
         if (input[i].decided) 
             printf("%d ", input[i].value);
         else
@@ -27,18 +27,18 @@ int main()
     run.push_back(Run(false, 0, Run::AT_LEAST));
 
     Automaton<bool> a(run);
-    const int MAX_LEN = 12;
+    const size_t MAX_LEN = 12;
     Input input[MAX_LEN];
 
-    for (int len = MAX_LEN/2; len <= MAX_LEN; len++) {
-        for (int i = 0; i < len; i++)
+    for (size_t len = MAX_LEN/2; len <= MAX_LEN; len++) {
+        for (size_t i = 0; i < len; i++)
             input[i] = Input(false, false);
         test(a, input, len);
     }
 
-    int len = MAX_LEN;
-    for (int pos = 0; pos < len; pos++) {
-        for (int i = 0; i < len; i++)
+    size_t len = MAX_LEN;
+    for (size_t pos = 0; pos < len; pos++) {
+        for (size_t i = 0; i < len; i++)
             input[i] = Input(false, false);
         input[pos].value   = true;
         input[pos].decided = true;
@@ -46,8 +46,8 @@ int main()
     }
 
     len = MAX_LEN-1;
-    for (int pos = 0; pos < len; pos++) {
-        for (int i = 0; i < len; i++)
+    for (size_t pos = 0; pos < len; pos++) {
+        for (size_t i = 0; i < len; i++)
             input[i] = Input(false, false);
         input[pos].value   = false;
         input[pos].decided = true;
