+++ Same.h
@@ -22,15 +22,15 @@ bool Same<T>::OnDecided(Variable<T> *decided)
 
     // Apply the same value to other variables' domains
     vector<Variable<T> *>  &variables = Constraint<T>::variables;
-    int num_variables = variables.size();
-    int value = decided->GetValue(0);
+    size_t num_variables = variables.size();
+    size_t value = decided->GetValue(0);
 
-    for (int i = 0; i < num_variables; i++) {
+    for (size_t i = 0; i < num_variables; i++) {
         if (variables[i] == decided) continue;
 
-        int old_domain_size = variables[i]->GetDomainSize();
+        size_t old_domain_size = variables[i]->GetDomainSize();
         variables[i]->Decide(value);
-        int new_domain_size = variables[i]->GetDomainSize();
+        size_t new_domain_size = variables[i]->GetDomainSize();
 
         if (new_domain_size == 0)
             return false;
